Room, LiveData, ViewModel functional on Source

+++ app/src/main/java/biz/eventually/atpl/data/network/SourceNetwork.kt
@@ -3,4 +3,4 @@ package biz.eventually.atpl.data.network
 /**
  * Created by thibault on 21/03/17.
  */
-data class SourceNetwork(val id: Int, val name: String)
+data class SourceNetwork(val id: Long, val name: String)
